Not working, but first clean compile
So this is definitely a milestone in so far that the bulk of the code changes to convert the firmware from monolithic mode to modular mode have been completed, and it compiles.
What this *does* *not* mean is that anything works, or that this is the final general configuration.  Already the "Gate" class and the "Signal" class are being reviewed in the "fitness for purpose" perspective.  For Gate this is really the realisation that this is just another name for Signal.  For Signal there is a recognition that this *could* just be more useful if it was implemented as a byte queue; the ability to pass a flag or data with a "release" to do something could make some elements of the firmware much simpler and better able to support the function anticipated.
Time yet to review these considerations.