Internal Version 0.3.5 - working
This is a checkpoint of the firmware development indicating that this if the first set of "full rewrite" modules that result in a operation DCC Controller.  The level of debugging code and additional facilities (all optionally compiled as required) means that the general aim of keeping this on an Arduino UNO R3 / Nano R3 is currently missed (uses an Arduino Mega2560 at the moment).  Also note that the display update remains tardy but fully functional.  This too needs addressing.

Subsequent versions will be numbered from V0.4.0.